+++ app/src/main/kotlin/io/github/fobo66/wearmmr/ui/theme/Color.kt
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package theme
+package io.github.fobo66.wearmmr.ui.theme
 
 import androidx.compose.ui.graphics.Color
